+++ src/draft/se/jbee/inject/container/Prototron.java
@@ -1,3 +1,8 @@
+/*
+ *  Copyright (c) 2012-2013, Jan Bernitt 
+ *			
+ *  Licensed under the Apache License, Version 2.0, http://www.apache.org/licenses/LICENSE-2.0
+ */
 package se.jbee.inject.container;
 
 import se.jbee.inject.Parameter;